rename DeckTab to DecksTab
